refactor(policy-creator): extract cmdk filter; bind palette listener once
Two improvements to the Playground command palette:

  M2 — keydown listener was being torn down and re-installed on
       every open/close because the effect depended on `open`. That
       briefly leaves the global capture-phase listener un-installed
       and forces Fumadocs' default-phase ⌘K handler to win during
       that window. Switch to a functional setOpen so we can read
       the previous state inside the callback and bind the listener
       exactly once for the component lifetime.

  Extracted filterIndex + IndexEntry into a pure helper file
  (playground/cmdk-filter.ts). The matcher is the part most worth
  testing (token AND-match, label-vs-keyword scoring), and pulling
  it out of a 'use client' JSX module lets node:test import it
  directly. command-palette.tsx re-exports filterIndex so existing
  call sites stay stable.
